Fix eslint rules and refactoring
