+++ src/screens/Home/index.tsx
@@ -1,19 +1,20 @@
 import React from 'react';
 import styled from 'styled-components';
 import { StatusBar } from 'expo-status-bar';
-import { SafeAreaView } from 'react-native';
+import { SafeAreaView, Text } from 'react-native';
 
 const Container = styled(SafeAreaView)`
   flex: 1;
   width: 100%;
   height: 100%;
-  justifyContent: center;
-  alignItems: center;
+  justify-content: center;
+  align-items: center;
 `;
 
 const Home = (): JSX.Element => (
   <Container>
     <StatusBar style="auto" />
+    <Text>Home</Text>
   </Container>
 );
